// src/lib/types.ts
export interface UserProfile {
    id: string;
    uid: string;
    displayName: string;
    email: string;
    role: string;
    photoURL: string;
    status?: 'Invite sent' | 'User Registered' | 'Pending';
    permissions?: string[];
}